Refactor EntityTickCallback into it's own module
