+++ fabric-events-entity-v1/src/main/resources/fabric.mod.json
@@ -1,0 +1,26 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-events-entity-v1",
+  "name": "Fabric Events Entity (v1)",
+  "version": "${version}",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-events-entity-v1/icon.png",
+  "contact": {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "depends": {
+    "fabricloader": ">=0.4.0",
+    "fabric-api-base": "*"
+  },
+  "description": "Events related to entities.",
+  "mixins": [
+    "fabric-events-entity-v1.mixins.json"
+  ]
+}
