+++ settings.gradle
@@ -22,6 +22,7 @@ include 'fabric-command-api-v1'
 include 'fabric-containers-v0'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
+include 'fabric-events-entity-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-events-lifecycle-v0'
 include 'fabric-item-groups-v0'
